+++ demetra-dotstat-desktop/src/main/java/be/nbb/demetra/sdmx/web/SdmxWebConfiguration.java
@@ -26,7 +26,9 @@ import java.io.File;
 import java.io.IOException;
 import java.util.Locale;
 import java.util.Properties;
+import java.util.logging.Level;
 
+@lombok.extern.java.Log
 @lombok.Data
 public class SdmxWebConfiguration {
 
@@ -62,6 +64,10 @@ public class SdmxWebConfiguration {
     private static final boolean DEFAULT_DISPLAY_CODES = false;
     private boolean displayCodes = DEFAULT_DISPLAY_CODES;
 
+    private static final String LOG_EVENTS_PROPERTY = "logEvents";
+    private static final boolean DEFAULT_LOG_EVENTS = false;
+    private boolean logEvents = DEFAULT_LOG_EVENTS;
+
     public static SdmxWebConfiguration copyOf(SdmxWebConfiguration bean) {
         SdmxWebConfiguration result = new SdmxWebConfiguration();
         result.sources = bean.sources;
@@ -72,6 +78,7 @@ public class SdmxWebConfiguration {
         result.noDefaultSSL = bean.noDefaultSSL;
         result.noSystemSSL = bean.noSystemSSL;
         result.displayCodes = bean.displayCodes;
+        result.logEvents = bean.logEvents;
         return result;
     }
 
@@ -99,10 +106,18 @@ public class SdmxWebConfiguration {
 
     private void reportEvent(WebSource source, String marker, CharSequence message) {
         StatusDisplayer.getDefault().setStatusText(message.toString());
+        if (logEvents) {
+            log.log(Level.INFO, () -> asLogMessage(source, marker, message));
+        }
     }
 
     private void reportError(WebSource source, String marker, CharSequence message, IOException error) {
         NotificationDisplayer.getDefault().notify(message.toString(), SdmxIcons.getDefaultIcon(), "", null);
+        log.log(Level.SEVERE, error, () -> asLogMessage(source, marker, message));
+    }
+
+    private static String asLogMessage(WebSource source, String marker, CharSequence message) {
+        return "[" + source.getId() + "] (" + marker + ") " + message;
     }
 
     Sheet toSheet() {
@@ -151,6 +166,11 @@ public class SdmxWebConfiguration {
                 .display("No system SSL")
                 .description("Disable system truststore")
                 .add();
+        b.withBoolean()
+                .select(this, LOG_EVENTS_PROPERTY)
+                .display("Log events")
+                .description("Log events in IDE logs")
+                .add();
         result.put(b.build());
 
         return result;
@@ -186,6 +206,7 @@ public class SdmxWebConfiguration {
         private final IParam<Config, Toggle> noDefaultSSL = Params.onEnum(DEFAULT_NO_DEFAULT_SSL, NO_DEFAULT_SSL_PROPERTY);
         private final IParam<Config, Toggle> noSystemSSL = Params.onEnum(DEFAULT_NO_SYSTEM_SSL, NO_SYSTEM_SSL_PROPERTY);
         private final IParam<Config, Boolean> displayCodes = Params.onBoolean(DEFAULT_DISPLAY_CODES, DISPLAY_CODES_PROPERTY);
+        private final IParam<Config, Boolean> logEvents = Params.onBoolean(DEFAULT_LOG_EVENTS, LOG_EVENTS_PROPERTY);
 
         @Override
         protected Config doForward(SdmxWebConfiguration a) {
@@ -198,6 +219,7 @@ public class SdmxWebConfiguration {
             noDefaultSSL.set(result, a.getNoDefaultSSL());
             noSystemSSL.set(result, a.getNoSystemSSL());
             displayCodes.set(result, a.isDisplayCodes());
+            logEvents.set(result, a.isLogEvents());
             return result.build();
         }
 
@@ -212,6 +234,7 @@ public class SdmxWebConfiguration {
             result.setNoDefaultSSL(noDefaultSSL.get(b));
             result.setNoSystemSSL(noSystemSSL.get(b));
             result.setDisplayCodes(displayCodes.get(b));
+            result.setLogEvents(logEvents.get(b));
             return result;
         }
     }